+++ src/driver/driver.service.ts
@@ -2,11 +2,11 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { DriverEntity } from './entity/driver.entity';
-//import { BusinessLogicException, BusinessError } from '../shared/errors/businessLogic-errors';
+import { BusinessLogicException, BusinessLogicError} from '../shared/errors/businessLogicError';
 
 
 @Injectable()
-export class BusinessService{
+export class DriverService{
     // Constructor
     constructor(
         @InjectRepository(DriverEntity)
@@ -16,7 +16,7 @@ export class BusinessService{
     // Functions
 
     // Create
-    async createBusiness(business: DriverEntity): Promise<DriverEntity>{
+    async create(business: DriverEntity): Promise<DriverEntity>{
         return await this.driverRepository.save(business)
     }
 
@@ -29,17 +29,17 @@ export class BusinessService{
     async findOne(id: string): Promise<DriverEntity>{
         const driver: DriverEntity = await this.driverRepository.findOne({where:{id}});
         // Manage error with class 
-        //if(!business)
-        //    throw new BusinessLogicException("The business with the given id was not found", BusinessError.NOT_FOUND);
+        if(!driver)
+            throw new BusinessLogicException("The driver with the given id was not found", BusinessLogicError.NOT_FOUND);
         return driver
     }
 
     // Update
     async update(id: string, driver: DriverEntity): Promise<DriverEntity> {
-        const persistedBusiness: DriverEntity = await this.driverRepository.findOne({where:{id}});
+        const persisteddriver: DriverEntity = await this.driverRepository.findOne({where:{id}});
         // Manage error with class 
-        //if (!persistedBusiness)
-        //  throw new BusinessLogicException("The business with the given id was not found", BusinessError.NOT_FOUND);
+        if (!persisteddriver)
+          throw new BusinessLogicException("The driver with the given id was not found", BusinessLogicError.NOT_FOUND);
         driver.id = id; 
         return await this.driverRepository.save(driver);
     }
@@ -48,8 +48,8 @@ export class BusinessService{
     async delete(id: string) {
         const driver: DriverEntity = await this.driverRepository.findOne({where:{id}});
         // Manage error with class 
-        //if (!business)
-        //  throw new BusinessLogicException("The business with the given id was not found", BusinessError.NOT_FOUND);
+        if (!driver)
+          throw new BusinessLogicException("The driver with the given id was not found", BusinessLogicError.NOT_FOUND);
         await this.driverRepository.remove(driver);
     }
 }
