+++ src/payment/payment.service.ts
@@ -2,11 +2,11 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { PaymentEntity } from './entity/payment.entity';
-//import { BusinessLogicException, BusinessError } from '../shared/errors/businessLogic-errors';
+import { BusinessLogicException, BusinessLogicError} from '../shared/errors/businessLogicError';
 
 
 @Injectable()
-export class BusinessService{
+export class PaymentService{
     // Constructor
     constructor(
         @InjectRepository(PaymentEntity)
@@ -16,8 +16,8 @@ export class BusinessService{
     // Functions
 
     // Create
-    async createBusiness(business: PaymentEntity): Promise<PaymentEntity>{
-        return await this.paymentRepository.save(business)
+    async create(payment: PaymentEntity): Promise<PaymentEntity>{
+        return await this.paymentRepository.save(payment)
     }
 
     // Get all
@@ -29,17 +29,17 @@ export class BusinessService{
     async findOne(id: string): Promise<PaymentEntity>{
         const payment: PaymentEntity = await this.paymentRepository.findOne({where:{id}});
         // Manage error with class 
-        //if(!business)
-        //    throw new BusinessLogicException("The business with the given id was not found", BusinessError.NOT_FOUND);
+        if(!payment)
+            throw new BusinessLogicException("The payment with the given id was not found", BusinessLogicError.NOT_FOUND);
         return payment
     }
 
     // Update
     async update(id: string, payment: PaymentEntity): Promise<PaymentEntity> {
-        const persistedBusiness: PaymentEntity = await this.paymentRepository.findOne({where:{id}});
+        const persistedPayment: PaymentEntity = await this.paymentRepository.findOne({where:{id}});
         // Manage error with class 
-        //if (!persistedBusiness)
-        //  throw new BusinessLogicException("The business with the given id was not found", BusinessError.NOT_FOUND);
+        if (!persistedPayment)
+          throw new BusinessLogicException("The payment with the given id was not found", BusinessLogicError.NOT_FOUND);
         payment.id = id; 
         return await this.paymentRepository.save(payment);
     }
@@ -48,8 +48,8 @@ export class BusinessService{
     async delete(id: string) {
         const payment: PaymentEntity = await this.paymentRepository.findOne({where:{id}});
         // Manage error with class 
-        //if (!business)
-        //  throw new BusinessLogicException("The business with the given id was not found", BusinessError.NOT_FOUND);
+        if (!payment)
+          throw new BusinessLogicException("The payment with the given id was not found", BusinessLogicError.NOT_FOUND);
         await this.paymentRepository.remove(payment);
     }
 }
